refactor(Lotto): Classfiy controller
랜덤으로 로또 생성해주는 메소드 이전.
Lotto -> LottoController
